+++ src/Resources/contao/dca/tl_vehicle_account.php
@@ -57,6 +57,7 @@ $GLOBALS['TL_DCA'][$strTable] = [
                 'icon' => 'delete.svg',
             ],
             'toggle' => [
+                'href' => 'act=toggle&amp;field=enabled',
                 'label' => &$GLOBALS['TL_LANG'][$strTable]['enable'],
                 'attributes' => 'onclick="Backend.getScrollOffset();return AjaxRequest.toggleVisibility(this,%s)"',
                 'icon' => 'visible.svg',
@@ -108,7 +109,7 @@ $GLOBALS['TL_DCA'][$strTable] = [
             'exclude' => true,
             'inputType' => 'select',
             'options' => &$GLOBALS['TL_LANG'][$strTable]['apiTypeOptions'],
-            'eval' => ['includeBlankOption' => true, 'tl_class' => 'w50'],
+            'eval' => ['includeBlankOption' => true, 'tl_class' => 'w50', 'submitOnChange' => true],
             'sql' => "varchar(64) NOT NULL default ''",
         ],
         'api_explanation' => [
@@ -148,11 +149,11 @@ $GLOBALS['TL_DCA'][$strTable] = [
             'sql' => "varchar(64) NOT NULL default ''",
         ],
         'enabled' => [
-            'exclude' => true,
-            'default' => false,
+            'toggle' => true,
+            'filter' => true,
             'inputType' => 'checkbox',
-            'eval' => ['doNotCopy'=>true],
-            'sql' => ['type' => 'boolean', 'default' => false],
+            'eval' => ['doNotCopy' => true],
+            'sql' => ['type' => 'boolean', 'default' => false]
         ],
     ],
 ];